+++ core/domain/src/main/java/com/hegunhee/domain/repository/twitch/TwitchSearchStreamerRepository.kt
@@ -1,0 +1,11 @@
+package com.hegunhee.domain.repository.twitch
+
+import com.hegunhee.domain.model.twitch.SearchData
+
+interface TwitchSearchStreamerRepository {
+
+    suspend fun getSearchStreamerData(streamerId: String): Result<SearchData>
+
+    suspend fun getSearchStreamerDataList(streamerId: String): Result<List<SearchData>>
+
+}
