+++ core/domain/src/main/java/com/hegunhee/domain/usecase/twitch/GetSearchStreamerDataListUseCase.kt
@@ -2,11 +2,12 @@ package com.hegunhee.domain.usecase.twitch
 
 import com.hegunhee.domain.model.twitch.SearchData
 import com.hegunhee.domain.repository.Repository
+import com.hegunhee.domain.repository.twitch.TwitchSearchStreamerRepository
 import javax.inject.Inject
 
-class GetSearchStreamerDataListUseCase @Inject constructor(private val repository : Repository) {
+class GetSearchStreamerDataListUseCase @Inject constructor(private val twitchSearchStreamerRepository: TwitchSearchStreamerRepository) {
 
     suspend operator fun invoke(streamerName : String) : Result<List<SearchData>> {
-        return repository.getSearchStreamerDataList(streamerName)
+        return twitchSearchStreamerRepository.getSearchStreamerDataList(streamerName)
     }
 }
