+++ core/domain/src/main/java/com/hegunhee/domain/usecase/twitch/GetSearchStreamerDataUseCase.kt
@@ -1,12 +1,12 @@
 package com.hegunhee.domain.usecase.twitch
 
 import com.hegunhee.domain.model.twitch.SearchData
-import com.hegunhee.domain.repository.Repository
+import com.hegunhee.domain.repository.twitch.TwitchSearchStreamerRepository
 import javax.inject.Inject
 
-class GetSearchStreamerDataUseCase @Inject constructor(private val repository: Repository) {
+class GetSearchStreamerDataUseCase @Inject constructor(private val twitchSearchStreamerRepository: TwitchSearchStreamerRepository) {
 
     suspend operator fun invoke(streamerId : String) : Result<SearchData> {
-        return repository.getSearchStreamerData(streamerId)
+        return twitchSearchStreamerRepository.getSearchStreamerData(streamerId)
     }
 }
